@@ -1,28 +1,28 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
-import { SocialMediaComponent } from './components/social-media/social-media.component';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { AboutComponent } from './components/about/about.component';
-import { FooterComponent } from './components/footer/footer.component';
-import { FormModule } from './components/form/form.module';
 import { HttpClientModule } from '@angular/common/http';
-
+import { FooterComponent } from './shared/footer/footer.component';
+import { NavbarComponent } from './shared/navbar/navbar.component';
+import { SocialMediaComponent } from './shared/social-media/social-media.component';
+import { SubmitToSheetsModule } from './modules/submit-to-sheets/submit-to-sheets.module';
+import { ToastrModule } from 'ngx-toastr';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    SocialMediaComponent,
     NavbarComponent,
-    AboutComponent,
     FooterComponent,
-    
+    SocialMediaComponent
   ],
   imports: [
     BrowserModule,
-    FormModule,
-    HttpClientModule
+    BrowserAnimationsModule,
+    HttpClientModule,
+    SubmitToSheetsModule,
+    ToastrModule.forRoot(),
   ],
   providers: [],
   bootstrap: [AppComponent]
