@@ -1,14 +1,18 @@
 import { Component, OnInit } from '@angular/core';
 import {
   AbstractControlOptions,
-  FormBuilder,
-  FormGroup,
+  UntypedFormBuilder,
+  UntypedFormGroup,
   Validators,
 } from '@angular/forms';
 import { distinctUntilChanged, empty, switchMap, tap } from 'rxjs';
 import { ValidatorField } from 'src/app/helpers/validator/ValidatorField';
 import { ViacepService } from 'src/app/services/viacep.service';
 import { AddressViewModel } from 'src/app/view-models/address.view-model';
+import { SheetsService } from '../../services/sheets.service';
+import { FormsViewModel } from '../../view-models/forms.view-model';
+import { ToastrService } from 'ngx-toastr';
+import { HttpStatusCode } from '@angular/common/http';
 
 @Component({
   selector: 'app-form',
@@ -16,9 +20,14 @@ import { AddressViewModel } from 'src/app/view-models/address.view-model';
   styleUrls: ['./form.component.css'],
 })
 export class FormComponent implements OnInit {
-  form!: FormGroup;
+  form!: UntypedFormGroup;
 
-  constructor(private formBuilder: FormBuilder, private viacepService : ViacepService) {}
+  constructor(
+    private formBuilder: UntypedFormBuilder,
+    private viacepService: ViacepService,
+    private sheetsService: SheetsService,
+    private toastr: ToastrService
+  ) {}
 
   ngOnInit(): void {
     this.validation();
@@ -28,7 +37,6 @@ export class FormComponent implements OnInit {
     return this.form.controls;
   }
 
- 
   private validation(): void {
     const formOptions: AbstractControlOptions = {
       validators: [
@@ -40,7 +48,7 @@ export class FormComponent implements OnInit {
         ValidatorField.MuchYearsOldAge('age'),
         ValidatorField.ValidCPF('cpf'),
         ValidatorField.SignNotExists('sign'),
-        ValidatorField.MustBeStrong('password'),       
+        ValidatorField.MustBeStrong('password'),
       ],
     };
     this.form = this.formBuilder.group(
@@ -57,12 +65,12 @@ export class FormComponent implements OnInit {
         email: ['', [Validators.required, Validators.email]],
         password: ['', [Validators.required]],
         confirmPassword: ['', Validators.required],
-        cep: ['', Validators.required,], 
+        cep: ['', Validators.required],
         address: ['', Validators.required],
         number: ['', Validators.required],
         district: ['', Validators.required],
         city: ['', Validators.required],
-        phone: ['', Validators.required ],
+        phone: ['', Validators.required],
         height: ['', Validators.required],
         weight: ['', Validators.required],
         blood: ['', Validators.required],
@@ -70,57 +78,70 @@ export class FormComponent implements OnInit {
       },
       formOptions
     );
-    this.CEPValidator()
-    
+    this.CEPValidator();
   }
 
-  CEPValidator(){
-    this.form.get('cep')?.statusChanges
-    .pipe(
-      distinctUntilChanged(),
-      switchMap(status => status === 'VALID' ?
-        this.viacepService.queryCEP(this.form.get('cep')?.value)
-        : empty()
+  CEPValidator() {
+    this.form
+      .get('cep')
+      ?.statusChanges.pipe(
+        distinctUntilChanged(),
+        switchMap((status) =>
+          status === 'VALID'
+            ? this.viacepService.queryCEP(this.form.get('cep')?.value)
+            : empty()
+        )
       )
-    )
-    .subscribe(dados  => dados ? this.AutoCompleteCEP(dados) : {});
+      .subscribe((dados) => (dados ? this.AutoCompleteCEP(dados) : {}));
   }
 
-  AutoCompleteCEP(dados : AddressViewModel){
+  AutoCompleteCEP(dados: AddressViewModel) {
     this.form.patchValue({
       address: dados.logradouro,
       district: dados.bairro,
       city: dados.localidade,
-    })
+    });
   }
 
   submitToSheets() {
-    const formObj = this.form.getRawValue();
-    console.log(formObj);
+    if(this.form.invalid){
+      this.toastr.error('YOU NEED TO INSERT ALL FIELDS CORRECTLY', 'ERROR' );
+      return;
+    }
+
+      const formObj = this.form.getRawValue();
+      this.sheetsService.writeOnSheet(formObj as FormsViewModel)
+    .subscribe(res => {
+        if( res.statusCode == HttpStatusCode.Created)
+        this.toastr.success('ROW ADDED', 'SUCCESS' );
+         else
+        this.toastr.error('CANT ADD ROW TO SHEETS, TRY AGAIN!', 'ERROR' );
+    })
+
   }
-
-  setAge(event : any){
-    const date = this.f.birthdate
-    if(!date.errors){
-      
+       
+  setAge(event: any) {
+    const date = this.f.birthdate;
+    if (!date.errors) {
       const birthdate = new Date(date.value);
-      let age = this.getAge(birthdate)
-       this.form.controls['age'].setValue(age)
-
-    }else{
-      this.form.controls['age'].setValue(null)
+      let age = this.getAge(birthdate);
+      this.form.controls['age'].setValue(age);
+    } else {
+      this.form.controls['age'].setValue(null);
     }
   }
-  getAge(birthdate : Date) {
+  getAge(birthdate: Date) {
     var today = new Date();
     var age = today.getFullYear() - birthdate.getFullYear();
     var m = today.getMonth() - birthdate.getMonth();
-    if (m < 0 || (m === 0 && today.getDate() -1 < birthdate.getDate())) {
-        age--;
+    if (m < 0 || (m === 0 && today.getDate() - 1 < birthdate.getDate())) {
+      age--;
     }
     return age;
-
   }
 
+  getData(){
 
+    this.toastr.warning('FEATURE NOT IMPLEMENTED', 'GENERATE DATA' );
+  }
 }
